Rename lambda to be called InitialiseSession
