+++ lambdas/initialisesession/src/main/resources/log4j2.yaml
